pyproject.toml
[tool.poetry]
name = "chioricord"
version = "0.1.0"
description = "Nulti-functional discord bot for your server"
authors = ["Milinuri Nirvalen <pentergust@mail.ru>"]
license = "GPL-3.0-or-later"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
loguru = "^0.7.2"
aiosqlite = "^0.20.0"
python-dotenv = "^1.0.1"
aiohttp = "^3.9.3"
discord-py = "^2.3.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.4"
isort = "^5.13.2"
icecream = "^2.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
